+++ src/components/card/card.css
@@ -1,0 +1,4 @@
+.textCont {
+	display: flex;
+	justify-content: center;
+}
