+++ src/utils/data.js
@@ -210,6 +210,6 @@ const databurg = [
 		"image_large": "https://code.s3.yandex.net/react/code/bun-01-large.png",
 		"__v": 0
 	}
-];
+]
 
 export { databurg };
